+++ packages/server/src/index.ts
@@ -1,6 +1,6 @@
 import { handle } from "hono/vercel";
 import { app } from "./base/base-app";
-import { authRoute } from "./routes/auth/register";
+import { authRoute } from "./routes/auth/auth";
 import { hc } from "hono/client";
 import { personalRoute } from "./routes/personal/me";
 import { secret } from "./routes/secret/test";
